# server/docs/swagger.yaml
definitions:
  connections.CreateConnectionRequest:
    properties:
      common_name:
        type: string
      db_name:
        type: string
      host:
        type: string
      password:
        type: string
      port:
        type: integer
      ssl_mode:
        type: string
      username:
        type: string
    type: object
  connections.UpdateConnectionRequest:
    properties:
      common_name:
        type: string
      database_name:
        type: string
      host:
        type: string
      id:
        type: integer
      port:
        type: integer
      ssl_mode:
        type: string
      username:
        type: string
    type: object
  interactDatabases.requestQuery:
    properties:
      query:
        type: string
    type: object
info:
  contact: {}
paths:
  /connections:
    get:
      description: Get all connections
      produces:
      - application/json
      responses:
        "200":
          description: Returns the connections
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Get all connections
      tags:
      - connections
  /connections/{id}:
    get:
      description: Get a connection by ID
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns the connection data
          schema:
            type: object
        "400":
          description: Returns that the connection ID is invalid
          schema:
            type: object
        "404":
          description: Returns that the connection does not exist
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Get a connection by ID
      tags:
      - connections
  /connections/create:
    post:
      consumes:
      - application/json
      description: Create a connection
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/connections.CreateConnectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the ID of the created connection
          schema:
            type: object
        "400":
          description: Returns that the request body is invalid or that some required
            fields are missing
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
        "502":
          description: If the connection could not be created
          schema:
            type: object
      summary: Create a connection
      tags:
      - connections
  /connections/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a connection
      parameters:
      - description: Connection ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns the success message
          schema:
            type: object
        "400":
          description: Returns that the connection ID is invalid
          schema:
            type: object
        "404":
          description: Returns that the connection does not exist
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
      summary: Delete a connection
      tags:
      - connections
  /ping:
    get:
      description: Ping the server
      produces:
      - application/json
      responses:
        "200":
          description: Returns pong
          schema:
            type: object
      summary: Ping the server
      tags:
      - general
  /query/{connectionId}:
    post:
      consumes:
      - application/json
      description: Query a database
      parameters:
      - description: Connection ID
        in: path
        name: connectionId
        required: true
        type: integer
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interactDatabases.requestQuery'
      produces:
      - application/json
      responses:
        "200":
          description: Returns the rows
          schema:
            type: object
        "400":
          description: If the request body is invalid
          schema:
            type: object
        "500":
          description: Internal error
          schema:
            type: object
        "501":
          description: If the connection could not be established
          schema:
            type: object
        "502":
          description: If the query could not be run
          schema:
            type: object
      summary: Query a database
      tags:
      - interactDatabases
  /schema/{connectionId}:
    get:
      consumes:
      - application/json
      description: Get the schema of a database
      parameters:
      - description: Connection ID
        in: path
        name: connectionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns the schema
          schema:
            type: object
        "400":
          description: Returns the error
          schema:
            type: object
        "501":
          description: If the connection could not be established or the schema could
            not be retrieved
          schema:
            type: object
      summary: Get the schema of a database
      tags:
      - interactDatabases
  /schema/refresh/{connectionId}:
    get:
      consumes:
      - application/json
      description: Refresh the schema of a database
      parameters:
      - description: Connection ID
        in: path
        name: connectionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Returns the schema
          schema:
            type: object
        "400":
          description: Returns the error
          schema:
            type: object
        "501":
          description: If the connection could not be established or the schema could
            not be retrieved
          schema:
            type: object
      summary: Refresh the schema of a database
      tags:
      - interactDatabases
swagger: "2.0"